services/content/models.py
"""Content Service — request/response models."""

from pydantic import BaseModel
from typing import Optional


class ContentUpload(BaseModel):
    title: str
    body: str
    content_type: Optional[str] = "lesson"
    metadata: Optional[dict] = None


class ContentSearch(BaseModel):
    query: str
    limit: Optional[int] = 5
